Delete bug fix.Add Integer id for pokemon dto

+++ app/src/main/java/com/pokemonify/pokemonify/pokemondatabase/PokemonDto.java
@@ -6,7 +6,7 @@ import java.io.Serializable;
  * Created by gaurav on 24/7/16.
  */
 public class PokemonDto implements Serializable {
-    private long id;
+    private int id;
     private String name;
     private int hp;
     private String imagePath = "-1";
@@ -21,7 +21,7 @@ public class PokemonDto implements Serializable {
     public PokemonDto() {
     }
 
-    public PokemonDto(long id, String name, int hp, String imagePath, String type, String desc, String adjectives, int weight, int height, int level) {
+    public PokemonDto(int id, String name, int hp, String imagePath, String type, String desc, String adjectives, int weight, int height, int level) {
         this.id = id;
         this.name = name;
         this.hp = hp;
@@ -34,11 +34,11 @@ public class PokemonDto implements Serializable {
         this.level = level;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
